latexml_engine

Macro DefMacro

macro_rules! DefMacro {
    ($proto:literal, None) => { ... };
    ($prototype:literal, sub [( $($var:ident),* )]
    $body:block $($input:tt)*) => { ... };
    ($proto:expr_2021, sub [$args:ident]
    $body:block $($input:tt)*) => { ... };
    ($proto:expr_2021, $body:block $($input:tt)*) => { ... };
    ($proto:literal, $expansion:literal $($input:tt)*) => { ... };
    ($cs:expr_2021, $parameters:literal, $expansion: literal) => { ... };
    ($cs:expr_2021, $parameters:literal, sub [$args:ident]
    $body:block $($input:tt)*) => { ... };
    ($cs:expr_2021, $parameters:literal, sub [( $($var:ident),* )]
    $body:block $($input:tt)*) => { ... };
    ($cs:expr_2021, $parameters:expr_2021, sub [$args:ident]
    $body:block $($input:tt)*) => { ... };
    ($cs:literal, None, $expansion:literal) => { ... };
    ($cs:literal, None, $expansion:literal, $($input:tt)*) => { ... };
    ($cs:literal, None, $expansion:expr_2021) => { ... };
    ($cs:expr_2021, None, $expansion:literal) => { ... };
    ($cs:expr_2021, None, $body:block) => { ... };
    ($cs:expr_2021, None, $expansion:expr_2021) => { ... };
    ($cs:expr_2021, None, $expansion:literal, $($input:tt)+) => { ... };
    ($cs:expr_2021, None, $expansion:expr_2021, $($input:tt)+) => { ... };
    ($cs:expr_2021, $parameters:expr_2021, $expansion:expr_2021) => { ... };
    ($proto:literal, None $($input:tt)*) => { ... };
    ($cs:expr_2021, $replacement:expr_2021, $expansion:expr_2021, $($input:tt)*) => { ... };
}

Defines a macro, a binding analog to \def

A prototype will be parsed into a command sequence and a list of parameters. Any macro arguments will be substituted for parameter indicators (eg #1) in the Tokens or tokenized string and the result is used as the expansion of the control sequence. If a closure is used, it is called at expansion time and should return a list of tokens as its result.