latexml_engine

Macro DefRegister

macro_rules! DefRegister {
    ($proto:expr_2021 => $value:expr_2021) => { ... };
    ($proto:expr_2021, $value:expr_2021) => { ... };
    ($cs:expr_2021, None, $value:expr_2021) => { ... };
    ($cs:expr_2021, None, $value:expr_2021, $($input:tt)+) => { ... };
    ($proto:expr_2021, $value:expr_2021, $($input:tt)+) => { ... };
    ($proto:expr_2021 => $value:expr_2021, $($input:tt)+) => { ... };
}

Defines a register with value as the initial value

(a Number, Dimension, Glue, MuGlue or Tokens — I haven’t handled Box’s yet). Usually, the prototype is just the control sequence, but registers are also handled by prototypes like \count{Number}. DefRegister arranges that the register value can be accessed when a numeric, dimension, … value is being read, and also defines the control sequence for assignment.