latexml_engine::prelude::binding::def

Module dialect

Structs

RegisterOptions

configuration for creating a new Register

Statics

ARG_HOLE

regex for the usual argument placeholders #1-#9

CONDITIONAL_CS_RE

regex for the prefix of a conditional command sequence

DIRTY_ID_IDIOM_RE

regex for ${}^{label}$

HAS_ARG_OR_CS

regex for a TeX argument specifier or any command sequence

LEADING_PROTOCOL_RE

regex for the prefix of a protocol (such as literal:)

NON_ID_CHARSET_RE

regex for characters not expected in a usual id attribute

SPACES_RE

regex for one-or-more spaces

TILDE_NOISE_RE

regex for a strange noisy TeX \\~{}

TRAILING_SLASH_RE

regex for a trailing slash (trivial, but aids replacement of said slash)

Functions

allocate_register

def_conditional

Define a conditional control sequence.

def_constructor

Binding definition connecting a TeX command sequence with a structured XML output.

def_environment

Defines an Environment that generates a specific XML fragment.

def_macro

Defines the macro expansion for a command sequence.

def_math

Define a Mathematical symbol or function.

def_math_constructor

Uses of DefMath without arguments, but with constructor-like options, are realized via a Constructor definition

def_math_dual

Advanced math replacements require a XMDual representation

def_math_primitive

EXPERIMENT: Introduce an intermediate case for simple symbols Define a primitive that will create a Tbox with the appropriate set of XMTok attributes.

def_primitive

Defines a primitive control sequence

def_register

Defines a register with an initial value.

dualize_arglist

Flesh out two dual (mathematical) forms of a given list of arguments.

get_xmarg_id

is_definable

Check if the token is not yet defined, or let to \relax

is_defined

Is defined in the LaTeX-y sense of also not being let to \relax.

is_defined_token

Token variant of is_defined. Defined in the LaTeX-y sense of also not being let to \relax.