latexml_engine::prelude::common

Module model

Structs

Model

TagFrame

Constants

LTX_NAMESPACE

Statics

MODEL

Functions

add_schema_declaration

add_tag_attribute

add_tag_content

can_contain

Can an element with (qualified name) tag contain a child element?

can_contain_sym

A check for allowed direct element containment, using ticket-based SymStr names.

can_have_attribute

decode_qname

Given a Qualified name, possibly prefixed with a namespace prefix, as defined by the code namespace mapping, return the NamespaceURI and localname.

decode_qname_sym

TODO: We need a proper data model to deal with the Symbol - String distinction. For now let’s allocate strings and release the arena, but this is a CODE SMELL!

get_document_namespace

get_document_namespace_prefix

get_document_namespace_prefixes

Returns all registered document namespace prefixes and their URIs.

get_namespace

get_namespace_prefix

Get the (code) prefix associated with $namespace, creating a dummy prefix and signalling an error if none has been registered.

get_node_document_qname

Same as get_node_qname, but using the Document namespace prefixes

get_node_qname

Get the node’s qualified name in standard form Ie. using the registered (code) prefix for that namespace. NOTE: Reconsider how Capture & WildCard should be integrated!?!

get_tag_contents

get_tags

initialize_model

is_in_schema_class

is_node_in_schema_class

is_permissive

load_schema

register_document_namespace

register_namespace

set_model

set_relaxng_schema

set_schema

set_schema_class

with_code_namespaces

with_node_qname

with_schema_data

Type Aliases

IndirectModel