latexml_engine::prelude::digested

Struct Digested

pub struct Digested(/* private fields */);

An Rc-guarded abstraction for any object encountered at the “digested” phase of processing

Implementations

impl Digested

pub fn data(&self) -> &DigestedData

immutably borrow the inner Digested data

pub fn value_of(&self) -> i64

Obtain the i64 value of the digested object, iff it wraps a RegisterValue

pub fn get_dimension(&self) -> Option<Dimension>

Obtain a Dimension from the digested object, iff it wraps a RegisterValue

pub fn pt_value(&self, prec: Option<u8>) -> f64

Obtain the f64 value of the digested object, iff it wraps a RegisterValue

pub fn any<F>(&self, check: F) -> bool
where F: FnMut(&Digested) -> bool,

Predicate check - true if any element of the current object passes the check

pub fn all<F>(&self, check: F) -> bool
where F: FnMut(&Digested) -> bool,

Predicate check - true if all elements of the current object passes the check

pub fn is_empty(&self) -> Result<bool, Error>

Predicate check - delegates to .is_empty() of the underlying data

pub fn is_skippable(&self) -> bool

Check if all items are “empty” or only spaces or otherwise skippable in a table cell. Perl: isSkippable (Alignment.pm L484-508)

pub fn raw_tokens(&self) -> Option<&Tokens>

Provide a way of emulating an Undigested argument, by requesting raw tokens, only when they are preserved – empty otherwise.

pub fn to_attribute(&self) -> String

builds an attribute-friendly String form of the digested object, suitable for XML attributes

pub fn untex(&self) -> Result<String, Error>

Reverts a digested object to Tokens and extracts a TeX-near string representation of its content

pub fn alignment_cell(&self) -> Option<&RefCell<Alignment>>