latexml_engine::prelude::parameter

Struct Parameter

pub struct Parameter {

    pub novalue: bool,
    pub semiverbatim: Option<Vec<char>>,
    pub optional: bool,
    pub name: SymbolU32,
    pub spec: SymbolU32,
    pub extra: Vec<Tokens>,
    pub inner: Option<Parameters>,
    pub reader: Rc<dyn Fn(Option<&Parameters>, &[Tokens]) -> Result<ArgWrap, Error>>,
    pub predigest: Option<Rc<dyn Fn(ArgWrap, &[Tokens]) -> Result<Option<Digested>, Error>>>,
    pub reversion: Option<Rc<dyn Fn(Vec<Token>, Option<&Parameters>, &[Tokens]) -> Result<Tokens, Error>>>,
    pub digested_reversion: Option<Rc<dyn Fn(&Digested) -> Result<Tokens, Error>>>,
    pub before_digest: Vec<Rc<dyn Fn() -> Result<Vec<Digested>, Error>>>,
    pub after_digest: Vec<Rc<dyn Fn(&mut Whatsit) -> Result<Vec<Digested>, Error>>>,

}

Fields
novalue: boolsemiverbatim: Option<Vec<char>>optional: boolname: SymbolU32spec: SymbolU32extra: Vec<Tokens>inner: Option<Parameters>reader: Rc<dyn Fn(Option<&Parameters>, &[Tokens]) -> Result<ArgWrap, Error>>predigest: Option<Rc<dyn Fn(ArgWrap, &[Tokens]) -> Result<Option<Digested>, Error>>>reversion: Option<Rc<dyn Fn(Vec<Token>, Option<&Parameters>, &[Tokens]) -> Result<Tokens, Error>>>digested_reversion: Option<Rc<dyn Fn(&Digested) -> Result<Tokens, Error>>>

Reversion closure that operates on the original Digested argument. Takes precedence over reversion when the argument is a digested value. Perl equivalent: reversion option on DefParameterType with undigested => 1.

before_digest: Vec<Rc<dyn Fn() -> Result<Vec<Digested>, Error>>>after_digest: Vec<Rc<dyn Fn(&mut Whatsit) -> Result<Vec<Digested>, Error>>>
Implementations

impl Parameter

pub fn new<T>( name: T, spec: T, extra: Option<Vec<Tokens>>, ) -> Result<Parameter, Error>
where T: AsRef<str>,

pub fn init(self) -> Result<Parameter, Error>

pub fn check_reader_function( name: &str, ) -> Option<Rc<dyn Fn(Option<&Parameters>, &[Tokens]) -> Result<ArgWrap, Error>>>

Obtain the reader of a given parameter name, if available

pub fn setup_catcodes(&self)

pub fn revert_catcodes(&self) -> Result<(), Error>

pub fn read(&self, fordefn: Option<&dyn Definition>) -> Result<ArgWrap, Error>

pub fn digest( &self, value_arg: ArgWrap, _fordefn: Option<&Constructor>, ) -> Result<Option<Digested>, Error>

pub fn revert(&self, value_opt: Option<Tokens>) -> Result<Option<Tokens>, Error>

pub fn reparse(&self, tokens: Tokens) -> Result<ArgWrap, Error>

This is needed by structured parameter types like KeyVals where the argument may already have been tokenized before the KeyVals (and the parameter types for the keys) had a chance to properly parse.